데이터베이스 연결 문자열 설정 방식 변경 및 네임스페이스 정리

+++ DotNetNote.SqlServer/dbo/Tables/Companies.sql
@@ -1,5 +1,5 @@
 ﻿CREATE TABLE [dbo].[Companies]
 (
-	[Id] INT NOT NULL PRIMARY KEY Identity(1, 1),
+	[Id] BIGINT NOT NULL PRIMARY KEY Identity(1, 1),
 	Name NVarChar(255) Not Null
 )
